Millix 189 (#59)
* MILLIX-189: convert /report-issue into send message form
