+++ src/config.js
@@ -2,11 +2,14 @@ export const WALLET_LOG_SIZE_MAX                       = 1000;
 export const GENESIS_TRANSACTION_ID                    = '2VngVznbdiQ5tqfWqn2NMP8DijqCbLX79Gygo9yYRVFU6iN35h';
 export const GENESIS_SHARD_ID                          = 'qGuUgMMVmaCvqrvoWG6zARjkrujGMpzJmpNhBgz1y3RjBG7ZR';
 export const CURRENCY_PAIR_SUMMARY_REFRESH_INTERVAL_MS = 10 * 60000;
+export const DISCORD_URL                               = 'https://discord.com/invite/nb5uaKq4yX';
+export const REPORT_ISSUE_ADDRESS                      = 'tcQAg1vsahrrdzA8KM3SmUqbt5KNV2pz1aVsqdPZ1kxw0b013NGMRi19V46EDABTn8vcy5Jg7NjefhYpJ';
 const config                                           = {
     WALLET_LOG_SIZE_MAX,
     GENESIS_TRANSACTION_ID,
     GENESIS_SHARD_ID,
-    CURRENCY_PAIR_SUMMARY_REFRESH_INTERVAL_MS
+    CURRENCY_PAIR_SUMMARY_REFRESH_INTERVAL_MS,
+    DISCORD_URL
 };
 
 export default config;
