+++ src/js/components/sidebar.jsx
@@ -138,7 +138,8 @@ class Sidebar extends Component {
         else if (section === 'help' &&
                  (
                      (defaultSelected === '/faq') ||
-                     (defaultSelected === '/report-issue')
+                     (defaultSelected === '/report-issue') ||
+                     (defaultSelected === '/system-info')
                  )
         ) {
             result = true;
@@ -427,6 +428,11 @@ class Sidebar extends Component {
                         <NavItem key={'report-issue'} eventKey="/report-issue">
                             <NavText>
                                 report problem
+                            </NavText>
+                        </NavItem>
+                        <NavItem key={'system-info'} eventKey="/system-info">
+                            <NavText>
+                                system info
                             </NavText>
                         </NavItem>
                     </NavItem>
